Add while loop and try..catch block around opening and reading from json file that contains status data.
This provides the ability for the web app code to catch the JSONDecode exception that will be thrown due to
the backend scheduled job writing to the file at the same the web page is being viewed. If that occurs then
it will wait for 3 seconds and then try again, in which the occurence of the return of the call will break
the 'while' loop. The end user should see minimal delay on the page loading unless there is a permanent reason
the the file doesn't contain the json data in the file.